# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users": {
      "post": {
        "operationId": "UsersController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "UsersController_getAllUsers",
        "parameters": [
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hasCompleted",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleMatch",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/v1/super-admin/directors": {
      "post": {
        "operationId": "SuperAdminController_createDirector",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Super Admin"
        ]
      },
      "get": {
        "operationId": "SuperAdminController_getAllDirectors",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Super Admin"
        ]
      }
    },
    "/api/v1/super-admin/directors/{id}": {
      "get": {
        "operationId": "SuperAdminController_getDirectorById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Super Admin"
        ]
      },
      "patch": {
        "operationId": "SuperAdminController_updateDirector",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Super Admin"
        ]
      },
      "delete": {
        "operationId": "SuperAdminController_deleteDirector",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Super Admin"
        ]
      }
    },
...
    "/api/v1/super-admin/users": {
      "get": {
        "operationId": "SuperAdminController_getAllUsers",
        "parameters": [
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Super Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "CCC Backend API",
    "description": "Christ Church Commission (CCC) Pastor Mentor platform REST API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT access token"
      }
    },
    "schemas": {
...
      "ObjectId": {
        "type": "object",
        "properties": {}
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 6
          },
          "role": {
            "type": "string",
            "enum": [
              "super admin",
              "director",
              "mentor",
              "field mentor",
              "pastor",
              "lay leader",
              "seminarian",
              "pending"
            ],
            "example": "pastor"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "accepted",
              "rejected"
            ],
            "example": "pending"
          },
          "interestId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "profilePicture": {
            "type": "string"
          },
          "isEmailVerified": {
            "type": "boolean"
          },
          "emailVerifiedAt": {
            "format": "date-time",
            "type": "string"
          },
          "isPasswordSet": {
            "type": "boolean"
          },
          "passwordCreatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "zoomUserId": {
            "type": "string"
          },
          "googleAccessToken": {
            "type": "string"
          },
          "googleRefreshToken": {
            "type": "string"
          },
          "googleTokenExpiry": {
            "type": "number"
          },
          "googleCalendarId": {
            "type": "string"
          },
          "googleCalendarStatus": {
            "type": "string",
            "enum": [
              "connected",
              "disconnected",
              "expired",
              "error"
            ]
          },
          "googleCalendarConnectedAt": {
            "format": "date-time",
            "type": "string"
          },
          "googleCalendarLastSyncAt": {
            "format": "date-time",
            "type": "string"
          },
          "googleCalendarEmail": {
            "type": "string"
          },
          "googleCalendarLastError": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email"
        ]
      },
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 6
          },
          "role": {
            "type": "string",
            "enum": [
              "super admin",
              "director",
              "mentor",
              "field mentor",
              "pastor",
              "lay leader",
              "seminarian",
              "pending"
            ],
            "example": "pastor"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "accepted",
              "rejected"
            ],
            "example": "pending"
          },
          "interestId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "profilePicture": {
            "type": "string"
          },
          "isEmailVerified": {
            "type": "boolean"
          },
          "emailVerifiedAt": {
            "format": "date-time",
            "type": "string"
          },
          "isPasswordSet": {
            "type": "boolean"
          },
          "passwordCreatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "zoomUserId": {
            "type": "string"
          },
          "googleAccessToken": {
            "type": "string"
          },
          "googleRefreshToken": {
            "type": "string"
          },
          "googleTokenExpiry": {
            "type": "number"
          },
          "googleCalendarId": {
            "type": "string"
          },
          "googleCalendarStatus": {
            "type": "string",
            "enum": [
              "connected",
              "disconnected",
              "expired",
              "error"
            ]
          },
          "googleCalendarConnectedAt": {
            "format": "date-time",
            "type": "string"
          },
          "googleCalendarLastSyncAt": {
            "format": "date-time",
            "type": "string"
          },
          "googleCalendarEmail": {
            "type": "string"
          },
          "googleCalendarLastError": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}